+++ src/Controller/LessonController.php
@@ -5,7 +5,6 @@ namespace App\Controller;
 use App\Entity\Course;
 use App\Entity\Lesson;
 use App\Form\LessonType;
-use App\Repository\LessonRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -18,20 +17,16 @@ use Symfony\Component\Routing\Annotation\Route;
 class LessonController extends AbstractController
 {
     /**
-     * @Route("/", name="lesson_index", methods={"GET"})
-     */
-    public function index(LessonRepository $lessonRepository): Response
-    {
-        return $this->render('lesson/index.html.twig', [
-            'lessons' => $lessonRepository->findAll(),
-        ]);
-    }
-
-    /**
      * @Route("/new/", name="lesson_new", methods={"GET","POST"})
      */
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(
+            'ROLE_SUPER_ADMIN',
+            $this->getUser(),
+            'У вас нет доступа к этой странице'
+        );
+
         $course_id = $request->query->get('course_id');
         $course = $entityManager->getRepository(Course::class)->find($course_id);
         $lesson = new Lesson();
@@ -58,6 +53,8 @@ class LessonController extends AbstractController
      */
     public function show(Lesson $lesson): Response
     {
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         return $this->render('lesson/show.html.twig', [
             'lesson' => $lesson,
         ]);
@@ -68,6 +65,12 @@ class LessonController extends AbstractController
      */
     public function edit(Request $request, Lesson $lesson): Response
     {
+        $this->denyAccessUnlessGranted(
+            'ROLE_SUPER_ADMIN',
+            $this->getUser(),
+            'У вас нет доступа к этой странице'
+        );
+
         $form = $this->createForm(LessonType::class, $lesson);
         $form->handleRequest($request);
 
@@ -88,6 +91,12 @@ class LessonController extends AbstractController
      */
     public function delete(Request $request, Lesson $lesson): Response
     {
+        $this->denyAccessUnlessGranted(
+            'ROLE_SUPER_ADMIN',
+            $this->getUser(),
+            'У вас нет доступа к этой странице'
+        );
+
         if ($this->isCsrfTokenValid('delete' . $lesson->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($lesson);
